Public API for getting information about build targets
Until now, SourceKit-LSP has been using a couple of internal data structures of the build plan, which were incidentally public, for this. This changes those data structures to internal and exposes a new public API for accessing information about build targets, including those for plugins.

Note that this API is part of a new module called `SwiftPM` which could be the sstart of a "proper" public API for SwiftPM. That said, right now it depends on several types of existing modules in its public interface and that isn't likely to be resolved as part of this particular change.

rdar://112120976